Support for multiple Steam library locations
Steam path search now checks each library for game and workshop
locations. This allows the game, the mod and workshops to be
installed in separate Steam libraries.

Library locations are pulled from the Steam `libraryfolders.vdf` file.
This file has undergone a few format changes in the past so the parsing
code is purposefully simple in the hope of maintaining compatibility.
In the file, any key named `path` is checked with only paths containing
the correct subdirectories used.